+++ cmd/fofa/main.go
@@ -9,7 +9,7 @@ import (
 )
 
 var (
-	version = "v0.2.23"
+	version = "v0.2.24"
 	commit  = "none"
 	date    = "unknown"
 	builtBy = "unknown" // goreleaser fill